fixed the test tolerance, refactor to the _test_solver_with_parameters function
The test_solver_tolerance function now tests that solvers based on tolerance can run with an appropriate n_iter (stop_val) based on the solver's strategy. This fixes the issue where it was previously testing the reverse of the expected behavior.

The code for looping through datasets to find one compatible with the solver has been refactored into the _test_solver_with_parameters function. This eliminates duplication with the code from test_solver